Add files via upload
A basic word counting script that accounts for potential errors in the user's entered file pathway.

The end game intention of this project is for me to be able to automatically track the number of words I write in a novel writing session automatically, without requiring any input from myself. I'm currently learning how to work with .docx files with the docx library, so for now, this is just a basic little program that counts the words in the file for me to iterate on in the future.